+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # `nonempty-collections`
 
+## Unreleased
+
+#### Added
+
+- Missing `FromNonEmptyIterator` for `HashMap`.
+
 ## 0.2.7 (2024-07-22)
 
 #### Added